backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1
